feat(types): separate types into different files
