+++ walledeval/types/__init__.py
@@ -1,59 +1,28 @@
 # walledeval/types/__init__.py
 
 from typing import Union
-from enum import Enum
 from pydantic import BaseModel
+
+from walledeval.types.llm import LLMType
+from walledeval.types.message import Message, Messages
+from walledeval.types.inputs import (
+    Prompt, Question,
+    AutocompleteTask,
+    MultipleChoiceQuestion,
+    MultipleResponseQuestion,
+    OpenEndedQuestion
+)
 
 __all__ = [
     "LLMType",
     "Message", "Messages",
+    "Prompt", "Question",
+    "AutocompleteTask",
     "MultipleChoiceQuestion",
     "MultipleResponseQuestion",
     "OpenEndedQuestion",
     "Log"
 ]
-
-
-class LLMType(Enum):
-    BASE = 0
-    INSTRUCT = 1
-    NEITHER = 2
-
-
-class Message(BaseModel):
-    role: str
-    content: str
-
-
-Messages = Union[
-    list[Message],
-    list[dict[str, str]],
-    str
-]
-
-
-class Prompt(BaseModel):
-    prompt: str
-
-
-class Question(BaseModel):
-    question: str
-
-
-class OpenEndedQuestion(Question):
-    pass
-
-
-class MultipleChoiceQuestion(Question):
-    question: str
-    choices: list[str]
-    answer: int = -1
-
-
-class MultipleResponseQuestion(Question):
-    question: str
-    choices: list[str]
-    answers: list[int] = []
 
 
 class Log(BaseModel):
